refactor: schedule 패키지를 calendar 로 변경
