+++ backend/src/main/java/com/woowacourse/momo/group/domain/calendar/Schedule.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.group.domain.schedule;
+package com.woowacourse.momo.group.domain.calendar;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
