+++ backend/src/test/java/com/woowacourse/momo/fixture/DurationFixture.java
@@ -5,7 +5,7 @@ import static com.woowacourse.momo.fixture.DateFixture.일주일후;
 
 import lombok.RequiredArgsConstructor;
 
-import com.woowacourse.momo.group.domain.schedule.Duration;
+import com.woowacourse.momo.group.domain.calendar.Duration;
 
 @SuppressWarnings("NonAsciiCharacters")
 @RequiredArgsConstructor
