+++ backend/src/test/java/com/woowacourse/momo/fixture/GroupFixture.java
@@ -15,8 +15,8 @@ import lombok.Getter;
 
 import com.woowacourse.momo.acceptance.group.GroupRestHandler;
 import com.woowacourse.momo.category.domain.Category;
-import com.woowacourse.momo.group.domain.schedule.Duration;
-import com.woowacourse.momo.group.domain.schedule.Schedule;
+import com.woowacourse.momo.group.domain.calendar.Duration;
+import com.woowacourse.momo.group.domain.calendar.Schedule;
 import com.woowacourse.momo.group.service.dto.response.GroupIdResponse;
 
 @SuppressWarnings("NonAsciiCharacters")
