+++ backend/src/test/java/com/woowacourse/momo/group/domain/calendar/DurationTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.group.domain.schedule;
+package com.woowacourse.momo.group.domain.calendar;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
